changing over to int64_t

+++ anuga/abstract_2d_finite_volumes/quantity.c
@@ -10,21 +10,22 @@
 // Ole Nielsen, GA 2004
 
 #include "math.h"
+#include "stdint.h"
 
 //Shared code snippets
 #include "util_ext.h"
 
-typedef long keyint;
+typedef int64_t keyint;
 
 //-------------------------------------------
 // Low level routines (called from wrappers)
 //------------------------------------------
 
-int _compute_gradients(keyint N,
+int64_t _compute_gradients(keyint N,
 			double* centroids,
 			double* centroid_values,
-			long* number_of_boundaries,
-			long* surrogate_neighbours,
+			int64_t* number_of_boundaries,
+			int64_t* surrogate_neighbours,
 			double* a,
 			double* b){
 
@@ -92,7 +93,7 @@ int _compute_gradients(keyint N,
 }
 
 
-int _compute_local_gradients(keyint N,
+int64_t _compute_local_gradients(keyint N,
 			       double* vertex_coordinates,
 			       double* vertex_values,
 			       double* a,
@@ -127,7 +128,7 @@ int _compute_local_gradients(keyint N,
     return 0;
 }
 
-int _extrapolate_from_gradient(keyint N,
+int64_t _extrapolate_from_gradient(keyint N,
 			       double* centroids,
 			       double* centroid_values,
 			       double* vertex_coordinates,
@@ -171,9 +172,9 @@ int _extrapolate_from_gradient(keyint N,
 }
 
 
-int _extrapolate_and_limit_from_gradient(keyint N,double beta,
+int64_t _extrapolate_and_limit_from_gradient(keyint N,double beta,
 					 double* centroids,
-					 long*   neighbours,
+					 int64_t*   neighbours,
 					 double* centroid_values,
 					 double* vertex_coordinates,
 					 double* vertex_values,
@@ -295,11 +296,11 @@ int _extrapolate_and_limit_from_gradient(keyint N,double beta,
 
 
 
-int _limit_vertices_by_all_neighbours(keyint N, double beta,
+int64_t _limit_vertices_by_all_neighbours(keyint N, double beta,
 				      double* centroid_values,
 				      double* vertex_values,
 				      double* edge_values,
-				      long*   neighbours,
+				      int64_t*   neighbours,
 				      double* x_gradient,
 				      double* y_gradient) {
 
@@ -362,11 +363,11 @@ int _limit_vertices_by_all_neighbours(keyint N, double beta,
 
 
 
-int _limit_edges_by_all_neighbours(keyint N, double beta,
+int64_t _limit_edges_by_all_neighbours(keyint N, double beta,
 				   double* centroid_values,
 				   double* vertex_values,
 				   double* edge_values,
-				   long*   neighbours,
+				   int64_t*   neighbours,
 				   double* x_gradient,
 				   double* y_gradient) {
 
@@ -448,11 +449,11 @@ int _limit_edges_by_all_neighbours(keyint N, double beta,
 }
 
 
-int _limit_edges_by_neighbour(keyint N, double beta,
+int64_t _limit_edges_by_neighbour(keyint N, double beta,
 		     double* centroid_values,
 		     double* vertex_values,
 		     double* edge_values,
-		     long*   neighbours) {
+		     int64_t*   neighbours) {
 
 	keyint i, k, k2, k3, k6;
 	keyint n;
@@ -503,13 +504,13 @@ int _limit_edges_by_neighbour(keyint N, double beta,
 }
 
 
-int _limit_gradient_by_neighbour(keyint N, double beta,
+int64_t _limit_gradient_by_neighbour(keyint N, double beta,
 		     double* centroid_values,
 		     double* vertex_values,
 		     double* edge_values,
 		     double* x_gradient,
 		     double* y_gradient,
-		     long*   neighbours) {
+		     int64_t*   neighbours) {
 
 	keyint i, k, k2, k3, k6;
 	keyint n;
@@ -559,7 +560,7 @@ int _limit_gradient_by_neighbour(keyint N, double beta,
 	return 0;
 }
 
-int _bound_vertices_below_by_constant(keyint N, double bound,
+int64_t _bound_vertices_below_by_constant(keyint N, double bound,
 		     double* centroid_values,
 		     double* vertex_values,
 		     double* edge_values,
@@ -610,7 +611,7 @@ int _bound_vertices_below_by_constant(keyint N, double bound,
 	return 0;
 }
 
-int _bound_vertices_below_by_quantity(keyint N,
+int64_t _bound_vertices_below_by_quantity(keyint N,
 				      double* bound_vertex_values,
 				      double* centroid_values,
 				      double* vertex_values,
@@ -661,7 +662,7 @@ int _bound_vertices_below_by_quantity(keyint N,
 	return 0;
 }
 
-int _interpolate(keyint N,
+int64_t _interpolate(keyint N,
 		 double* vertex_values,
 		 double* edge_values,
                  double* centroid_values) {
@@ -686,7 +687,7 @@ int _interpolate(keyint N,
 	return 0;
 }
 
-int _interpolate_from_vertices_to_edges(keyint N,
+int64_t _interpolate_from_vertices_to_edges(keyint N,
 					double* vertex_values,
 					double* edge_values) {
 
@@ -709,7 +710,7 @@ int _interpolate_from_vertices_to_edges(keyint N,
 }
 
 
-int _interpolate_from_edges_to_vertices(keyint N,
+int64_t _interpolate_from_edges_to_vertices(keyint N,
 					double* vertex_values,
 					double* edge_values) {
 
@@ -731,7 +732,7 @@ int _interpolate_from_edges_to_vertices(keyint N,
 	return 0;
 }
 
-int _backup_centroid_values(keyint N,
+int64_t _backup_centroid_values(keyint N,
 			    double* centroid_values,
 			    double* centroid_backup_values) {
     // Backup centroid values
@@ -748,7 +749,7 @@ int _backup_centroid_values(keyint N,
 }
 
 
-int _saxpy_centroid_values(keyint N,
+int64_t _saxpy_centroid_values(keyint N,
 			   double a,
 			   double b,
 			   double* centroid_values,
@@ -768,7 +769,7 @@ int _saxpy_centroid_values(keyint N,
 }
 
 
-int _update(keyint N,
+int64_t _update(keyint N,
 	    double timestep,
 	    double* centroid_values,
 	    double* explicit_update,
@@ -819,9 +820,9 @@ int _update(keyint N,
 }
 
 
-int _average_vertex_values(keyint N,
-			   long* vertex_value_indices,
-			   long* number_of_triangles_per_node,
+int64_t _average_vertex_values(keyint N,
+			   int64_t* vertex_value_indices,
+			   int64_t* number_of_triangles_per_node,
 			   double* vertex_values,
 			   double* A) {
   // Average vertex values to obtain one value per node
@@ -868,9 +869,9 @@ int _average_vertex_values(keyint N,
   return 0;
 }
 
-int _average_centroid_values(keyint N,
-			   long* vertex_value_indices,
-			   long* number_of_triangles_per_node,
+int64_t _average_centroid_values(keyint N,
+			   int64_t* vertex_value_indices,
+			   int64_t* number_of_triangles_per_node,
 			   double* centroid_values,
 			   double* A) {
   // Average centroid values to obtain one value per node
@@ -918,11 +919,11 @@ int _average_centroid_values(keyint N,
 // from a list of vertices and values at those vertices. Called in
 // quantity.py by _set_vertex_values.
 // Naming is a little confusing - but sticking with convention.
-int _set_vertex_values_c(keyint num_verts,
-                        long * vertices,
-                        long * node_index,
-                        long * number_of_triangles_per_node,
-                        long * vertex_value_indices,
+int64_t _set_vertex_values_c(keyint num_verts,
+                        int64_t * vertices,
+                        int64_t * node_index,
+                        int64_t * number_of_triangles_per_node,
+                        int64_t * vertex_value_indices,
                         double * vertex_values,
                         double * A
                         ){
@@ -945,10 +946,10 @@ int _set_vertex_values_c(keyint num_verts,
 
 }
 
-int _min_and_max_centroid_values(keyint N,
+int64_t _min_and_max_centroid_values(keyint N,
                                  double * qc,
                                  double * qv,
-                                 long * neighbours,
+                                 int64_t * neighbours,
                                  double * qmin,
                                  double * qmax){
